# Proposal of text pasting performance issue solution