@danielshahaf 2015-09-02
It would be interested to know whether slowness is still observed under zsh-5.1: that release added support for bracketed pasting, which change the way the line editor processes pasted characters.