@ferenczy 2015-09-14
OK, I'll try it as soon as zsh 5.1 is released for Cygwin (there is version 5.0.8 currently).