@danielshahaf 2015-11-17
At the zsh level, `$PENDING` is available on all systems that support the `FIONREAD` ioctl.  Can you check on your system, whether the preprocessor conditional block in zsh's `Src/Zle/utils.c:noquery()` is even compiled?  Perhaps the problem is that zsh on cygwin needs to do something to make that ioctl visible (include another header, for example)?

The relevant callsite is `zle_params.c:get_pending()`.