@danielshahaf 2016-06-05
Closing per https://github.com/zsh-users/zsh-syntax-highlighting/issues/153#issuecomment-222057200